python code check the number is neon number or not
This code follows an algorithm that checks whether the given number is a neon number or not. Here, we should be  known about a neon number before getting into this. A neon number is simply a number to which the sum of the digits of its square will be equal. So here we are checking whether the number and the sum of the digits its square is equal. If it is equal then it is a neon else it is not a neon number.