added algorithm "Conway's Game of Life"
Coded in both C language and Js(using HTML5 ) to have a web browser representation as well.
code in C : Prints a pattern a number of times as user input.
code in Js: allows you to click on the canvas to draw and erase cells and provides buttons to start/stop the simulation and clear the grid. The simulation will continue to evolve until you stop it, so that we can draw patterns as you like.