Create funnyalgo.cpp
This code initializes an array of jokes and then selects a random joke to display when you run the program. It uses the std::rand() function to generate a random index for the joke array and then displays the selected joke. Each time you run the program, you'll get a different joke.

Feel free to add more jokes to the jokes vector for even more variety!